+++ cypress/locators/duelling-picklist/locators.js
@@ -4,5 +4,3 @@ export const PICKLIST = '[data-element="picklist"]';
 export const PICKLIST_ITEMS = '[data-element="picklist-item"]';
 export const PICKLIST_LEFT_LABEL = '[data-element="picklist-left-label"]';
 export const PICKLIST_RIGHT_LABEL = '[data-element="picklist-right-label"]';
-export const ADD_ELEMENT = '[data-element="add"]';
-export const REMOVE_ELEMENT = '[data-element="remove"]';